# Draft: Split get_storage_at into two paths: one optimistic and simple, the other more comprehensive but slower.