Before:
```
# explain analyze select * from get_storage_at_by_hash('0x3a55e670852b2b985cd8c252fddae2acb4f798d0fecdcb1e2da53726332eb559', '0xceb09ad4b38c2b87567293b73d49fcdea74dadb2546fa8e55e157c9a302ed44f', '0x3a76d8d865efc3be160fc2fbb1925f36707b1a54daf5e61656e8cc0504797c7a');
                                                           QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------------
 Function Scan on get_storage_at_by_hash  (cost=0.25..10.25 rows=1000 width=77) (actual time=1234.447..1234.448 rows=1 loops=1)
 Planning Time: 0.055 ms
 Execution Time: 1234.472 ms
(3 rows)


# select * from get_storage_at_by_hash('0x3a55e670852b2b985cd8c252fddae2acb4f798d0fecdcb1e2da53726332eb559', '0xceb09ad4b38c2b87567293b73d49fcdea74dadb2546fa8e55e157c9a302ed44f', '0x3a76d8d865efc3be160fc2fbb1925f36707b1a54daf5e61656e8cc0504797c7a');
                              cid                              |                             mh_key                              | block_number | node_type | state_leaf_removed
---------------------------------------------------------------+-----------------------------------------------------------------+--------------+-----------+--------------------
 bagmacgza5pdieleejyn74xbye7kbxo4c77ehtbbprhqajg7d5ov4iuhuef5a | /blocks/DMQOXRUCFSCE4G76LQ4CPVA3XOBP7SDZQQXYTYAETPR6XK6EKD2CC6Q |     12499699 |         2 | f
(1 row)

```

After:

```
# explain analyze select * from test_get_storage_at_by_hash('0x3a55e670852b2b985cd8c252fddae2acb4f798d0fecdcb1e2da53726332eb559', '0xceb09ad4b38c2b87567293b73d49fcdea74dadb2546fa8e55e157c9a302ed44f', '0x3a76d8d865efc3be160fc2fbb1925f36707b1a54daf5e61656e8cc0504797c7a');
                                                           QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------
 Function Scan on test_get_storage_at_by_hash  (cost=0.25..10.25 rows=1000 width=77) (actual time=95.012..95.013 rows=1 loops=1)
 Planning Time: 0.049 ms
 Execution Time: 95.036 ms
(3 rows)

# select * from test_get_storage_at_by_hash('0x3a55e670852b2b985cd8c252fddae2acb4f798d0fecdcb1e2da53726332eb559', '0xceb09ad4b38c2b87567293b73d49fcdea74dadb2546fa8e55e157c9a302ed44f', '0x3a76d8d865efc3be160fc2fbb1925f36707b1a54daf5e61656e8cc0504797c7a');
                              cid                              |                             mh_key                              | block_number | node_type | state_leaf_removed
---------------------------------------------------------------+-----------------------------------------------------------------+--------------+-----------+--------------------
 bagmacgza5pdieleejyn74xbye7kbxo4c77ehtbbprhqajg7d5ov4iuhuef5a | /blocks/DMQOXRUCFSCE4G76LQ4CPVA3XOBP7SDZQQXYTYAETPR6XK6EKD2CC6Q |     12499699 |         2 | f
(1 row)
```